Applied some of the code review #29
