+++ src/custom/editors/CodeEditor.js
@@ -1,14 +1,12 @@
 import Editor from "@monaco-editor/react";
 import {useState} from "react";
 
-const MyEditor = ({language, data, setData, modelName, schema}) => {
+const CodeEditor = ({language, data, setData, modelName, schema}) => {
 
-    // TODO I get the heebie jeebies of variables starting with `my`, can you come up with more descriptive names?
-    // TODO same remark for this entire component :) Suggestion: CodeEditor
-    const [myEditor, setMyEditor] = useState(null);
-    const [myMonaco, setMyMonaco] = useState(null);
-    const [myModel, setMyModel] = useState(null)
-    const [myModelUri, setMyModelUri] = useState(null)
+    const [editorInstance, setEditorInstance] = useState(null);
+    const [monacoInstance, setMonacoInstance] = useState(null);
+    const [modelInstance, setModelInstance] = useState(null)
+    const [modelUriInstance, setModelUriInstance] = useState(null)
 
     function editorDidMountNodes(editor, monaco) {
         // https://microsoft.github.io/monaco-editor/playground.html#extending-language-services-configure-json-defaults
@@ -16,10 +14,10 @@ const MyEditor = ({language, data, setData, modelName, schema}) => {
         const modelUri = monaco.Uri.parse(modelName + ".json"); // a made up unique URI for our model
         const model = monaco.editor.createModel(data, 'json', modelUri);
 
-        setMyEditor(editor);
-        setMyMonaco(monaco);
-        setMyModel(model);
-        setMyModelUri(modelUri)
+        setEditorInstance(editor);
+        setMonacoInstance(monaco);
+        setModelInstance(model);
+        setModelUriInstance(modelUri)
 
         monaco.languages.json.jsonDefaults.setDiagnosticsOptions({
             validate: true,
@@ -40,18 +38,18 @@ const MyEditor = ({language, data, setData, modelName, schema}) => {
     // This component gets created multiple times and the changes you do on monaco or editor are probably overwritten
     // In this function, I just overwrite it again when you click with you mouse on the editor
     function initAgain() {
-        myMonaco.languages.json.jsonDefaults.setDiagnosticsOptions({
+        monacoInstance.languages.json.jsonDefaults.setDiagnosticsOptions({
             validate: true,
             schemas: [
                 {
                     // uri: "http://.....-schema.json",
-                    fileMatch: [myModelUri.toString()],
+                    fileMatch: [modelUriInstance.toString()],
                     schema: schema
                 }
             ]
         })
 
-        myEditor.setModel(myModel);
+        editorInstance.setModel(modelInstance);
     }
 
 
@@ -71,4 +69,4 @@ const MyEditor = ({language, data, setData, modelName, schema}) => {
     </>
 }
 
-export default MyEditor;
+export default CodeEditor;
