+++ src/custom/node/nodeUtil.js
@@ -4,6 +4,7 @@ import solid from "../../assets/solid.svg";
 
 // TODO rename this file to better cover its contents
 // TODO why not put these svgs as actual files and import them? Makes this code a lot cleaner + you can edit the svgs in dedicated programs more easily
+
 export function getShape(shapeId, fill, stroke, strokeWidth) {
     const SHAPES = {
         "8-star":
